Restructured test directory.
