+++ csunplugged/tests/BaseTestWithDB.py
@@ -35,7 +35,7 @@ class BaseTestWithDB(TestCase):
         '''Called before each test.
         Sets the language to English, creates a new user and logs into the database
         '''
-        activate('en')
+        activate('en') # TODO
         self.test_user = User.objects.create_user(self.username, self.email, self.password)
         login = self.client.login(username=self.username, password=self.password)
         self.assertEqual(login, True)
