+++ csunplugged/tests/topics/models/test_test_connected_generated_resource.py
