--- csunplugged/tests/topics/test_models.py
@@ -1,13 +1,0 @@
-# flake8: noqa
-from tests.topics.models.learning_outcome import LearningOutcomeModelTest
-from tests.topics.models.curriculum_area import CurriculumAreaModelTest
-from tests.topics.models.classroom_resource import ClassroomResourceModelTest
-from tests.topics.models.topic import TopicModelTest
-from tests.topics.models.unit_plan import UnitPlanModelTest
-from tests.topics.models.programming_exercise_difficulty import ProgrammingExerciseDifficultyModelTest
-from tests.topics.models.programming_exercise import ProgrammingExerciseModelTest
-from tests.topics.models.programming_exercise_language import ProgrammingExerciseLanguageModelTest
-from tests.topics.models.programming_exercise_language_implementation import ProgrammingExerciseLanguageImplementationModelTest
-from tests.topics.models.lesson import LessonModelTest
-from tests.topics.models.curriculum_integration import CurriculumIntegrationModelTest
-from tests.topics.models.connected_generated_resource import ConnectedGeneratedResourceModelTest
