--- csunplugged/tests/topics/test_urls.py
@@ -1,13 +1,0 @@
-# flake8: noqa
-from tests.topics.urls.index import IndexURLTest
-from tests.topics.urls.all_curriculum_integrations import AllCurriculumIntegrationsURLTest
-from tests.topics.urls.topic import TopicURLTest
-from tests.topics.urls.integration_list import IntegrationListURLTest
-from tests.topics.urls.integration import IntegrationURLTest
-from tests.topics.urls.other_resources import OtherResourcesURLTest
-from tests.topics.urls.unit_plan import UnitPlanURLTest
-from tests.topics.urls.lesson import LessonURLTest
-from tests.topics.urls.programming_exercises_list import ProgrammingExercisesListURLTest
-from tests.topics.urls.programming_exercise import ProgrammingExerciseURLTest
-from tests.topics.urls.programming_exercise_language_solution import ProgrammingExerciseLanguageSolutionURLTest
-from tests.topics.urls.programming_exercise_difficulty import ProgrammingExerciseDifficultyURLTest
