--- csunplugged/tests/topics/test_views.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from tests.topics.views.index_view import IndexViewTest
